% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_urls.R
\name{fetch_file_urls}
\alias{fetch_file_urls}
\title{Generate a tibble of URLs for data files}
\usage{
fetch_file_urls()
}
\value{
a tibble with three columns: type, year and file_url
}
\description{
Fetch the URLs of crime data files from the Crime Open Database server,
together with the type of data in the file and the year the data is for.
}
