% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crime_data.R
\name{get_crime_data}
\alias{get_crime_data}
\title{Get data from the Open Crime Database}
\usage{
get_crime_data(years = NULL, type = "sample", cache = TRUE)
}
\arguments{
\item{years}{A single number or vector of numbers specifying the years for
which data should be retrieved. If NULL (the default), data for all
available years will be returned.}

\item{type}{Either "sample" (the default), "core" or "extended".}

\item{cache}{Should the result be cached and then re-used if the function is
called again with the same arguments?}
}
\value{
A tibble containing data from the Open Crime Database.
}
\description{
Retrieves data from the Open Crime Database for the specified years.
}
\details{
By default this function returns a 1% sample of the 'core' data. This is the
default to minimise accidentally requesting large files over a network.

Seting type = "core" retrieves the core fields (e.g. the type, co-ordinates
and date/time of each offense) for each offense.
The data retrieved by setting type = "extended" includes all available fields
provided by the police department in each city. The extended data fields have
not been harmonized across cities, so will require further cleaning before
most types of analysis.

Compressed data files are approximately 40MB per year for the core data and
70MB per year for the extended data, so consider requesting fewer years of
data (or use type = "sample") for exploratory analysis.
}
\examples{
# Retrieve data for specific years
get_crime_data(years = 2011:2015, type = "sample")

}
