% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_urls.R
\name{get_file_urls}
\alias{get_file_urls}
\title{Get URLs for Open Crime Database files}
\usage{
get_file_urls(cache = TRUE, quiet = FALSE)
}
\arguments{
\item{cache}{Should the result be cached and then re-used if the function is
called again with the same arguments?}

\item{quiet}{Should messages and warnings relating to data availability be
suppressed?}
}
\description{
URLs are either obtained from the OSF API or, if a cached version exists,
from the cache.
}
