% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_urls.R
\name{list_crime_data}
\alias{list_crime_data}
\title{List Data Available in the Open Crime Database}
\usage{
list_crime_data(quiet = FALSE)
}
\arguments{
\item{quiet}{Should messages and warnings relating to data availability and
processing be suppressed?}
}
\value{
A tibble
}
\description{
Get a tibble showing what years of crime data are available from which cities
in the Open Crime Database.
}
\examples{
\donttest{
list_crime_data()
}

}
