% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimedata.R
\docType{package}
\name{crimedata}
\alias{crimedata}
\alias{crimedata-package}
\title{crimedata: a package for accessing US city crime data}
\description{
Access incident-level crime data from the Open Crime Database
}
\section{Crime Open Database}{


  The Crime Open Database (CODE) is a service that makes it convenient to use
  crime data from multiple US cities in research on crime. All the data are
  available to use for free as long as you acknowledge the source of the
  data.

  For more about CODE data, see \url{https://osf.io/zyaqn/}.
}

\section{Accessing the data}{


  To access CODE data, call \code{\link{get_crime_data}}. Data are returned
  as a 'tidy' tibble with each row corresponding to one recorded crime.
}

\section{Chicago data license}{


  This site provides applications using data that has been modified for use
  from its original source, \url{https://www.chicago.gov/}, the official
  website of the City of Chicago. The City of Chicago makes no claims as to
  the content, accuracy, timeliness, or completeness of any of the data
  provided at this site. The data provided at this site is subject to change
  at any time. It is understood that the data provided at this site is being
  used at one's own risk.
}

\seealso{
Useful links:
\itemize{
  \item \url{http://pkgs.lesscrime.info/crimedata/}
  \item \url{https://github.com/mpjashby/crimedata}
  \item Report bugs at \url{https://github.com/mpjashby/crimedata/issues}
}

}
\author{
\strong{Maintainer}: Matthew Ashby \email{matthew.ashby@ucl.ac.uk} (\href{https://orcid.org/0000-0003-4201-9239}{ORCID}) [copyright holder]

}
