% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crime_data.R
\name{get_crime_data}
\alias{get_crime_data}
\title{Get Data from the Open Crime Database}
\usage{
get_crime_data(years = NULL, cities = NULL, type = "sample",
  cache = TRUE, quiet = FALSE)
}
\arguments{
\item{years}{A single integer or vector of integers specifying the years for
which data should be retrieved. If NULL (the default), data for the most
recent year will be returned.}

\item{cities}{A character vector of city names for which data should be
retrieved. If NULL (the default), data for all available cities will be
returned.}

\item{type}{Either "sample" (the default), "core" or "extended".}

\item{cache}{Should the result be cached and then re-used if the function is
called again with the same arguments?}

\item{quiet}{Should messages and warnings relating to data availability and
processing be suppressed?}
}
\value{
A tibble containing data from the Open Crime Database.
}
\description{
Retrieves data from the Open Crime Database for the specified years.
}
\details{
By default this function returns a 1% sample of the 'core' data. This is the
default to minimise accidentally requesting large files over a network.

Seting type = "core" retrieves the core fields (e.g. the type, co-ordinates
and date/time of each offense) for each offense.
The data retrieved by setting type = "extended" includes all available fields
provided by the police department in each city. The extended data fields have
not been harmonized across cities, so will require further cleaning before
most types of analysis.

Requesting all data may lead to problems with memory capacity. Consider
downloading smaller quantities of data (e.g. using type = "sample") for
exploratory analysis.
}
\examples{
# Retrieve data for specific years
get_crime_data(years = 2011:2015, type = "sample")

}
