% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_urls.R
\name{get_file_urls}
\alias{get_file_urls}
\title{Get URLs for Open Crime Database files}
\usage{
get_file_urls()
}
\description{
URLs are either obtained from the OSF API or, if a cached version exists,
from the cache.
}
